rename: model/types to consts/types
