+++ src/c_widgets/LayoutHeader/ui/MiddleSlot/MiddleSlot.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames/bind';
 
 import { imgMap, Button } from '@/f_shared';
 
-import { MiddleSlotType } from '../../model/types';
+import { MiddleSlotType } from '../../consts/types';
 
 import css from './MiddleSlot.module.css';
